Ethernet Demo Driver for 85xx PCI Agent/EP.
This patch provide a virtual ethernet driver to demonstrate the 85xx
PCI Agent/EP funtion.

Two 85xx boards were need to demo the funciton, One act as Host/RC and the
other one act as Agent/EP(on one of their pci ports). The two boards were
connected by cable or the Agent/EP was pluged into the Host/RC's pci/pcie
slots.

The patch provide two modules, one for the Host/RC side and one for the
Agent/EP side. The Host/RC more likes a board and the Agent/EP more likes
a card, The module for Host/RC was named boardnet, and the one for Agent/EP
was named cardnet.

One mem page share memory was defined in the cardnet, the boardnet and the
cardnet exchange the packets and other informations through this share
memory, Two message interrupts in the cardnet were used to provide the
necessary interrupt between the two boards. One was used for the boardnet
to interrupt the cardnet, and the other one was routed to the
IRQ_OUT which will act as the interrupt from cardnet to boardnet. The
IRQ_OUT on the cardnet should connected to the boardnet on pci system. On
PCIE system, the IRQ_OUT can trig the MSI interrupt to the host side, So if
demo the PCI EP was needed, the CONFIG_PCI_MSI should be defined.

This is a demo application for 85xx BSP release and was tested on 8536DS
board, there are room for improvement in this driver, the buffer struct
for packet exchanging was very simple, only one packet can be sent out in one
interrupt and the packet exchange rate was restricted.
The share memory was set in the cardnet, so it was not symmetrical for the
cardnet and boardnet. The result is that it's much faster for the Host/RC
writing than for reading. We can try to set packet exchange buffer in both
boardnet and cardnet, but that may need to change a lot of settings for the
PCI host, which was not appropriate for the driver and limited the reusing
of the driver. Also as the share mem was set in cardnet, we need to
insmod the cardnet driver first(also need to bootup the Agent/EP first).

Signed-off-by: Jason Jin <Jason.jin@freescale.com>
Signed-off-by: Mingkai Hu <Mingkai.hu@freescale.com>