talitos: Adds SEC channel remap DTS support
This patch adds the support for remapping the SEC channels to
alternate region. The channel-remap field is a 4-bit bitmap in
which a value 0 represents no remapping and a value 1 means that
the channel is remapped to alternate region.
The bitmap is as follows:
0 - Channel 0
1 - Channel 1
2 - Channel 2
3 - Channel 3

A bitmap of 0x3 represents Channel 2 and Channel 3 being remapped.

Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>