gianfar: clearing the message interrupt by reading the message.
There seems to be issue if message interrupt is cleared by
clearing the status bit. After some time, one of the message
interrupt gets stuck and the cpu doesnt receive the interrupt.
The interrupt is in fact counted as bad interrupt.

This issue doesnt occur if we clear the interrupt by reading the
message. RM says that either of the two methods can be used.