gianfar: set required affinities of message interrupts
Message interrupts used in packet steering code requires
interrupt affinities to be set correctly for proper functioning.
If the interrupt affinities are not correctly set, traffic will
not flow and also the system can hang. Hence, setting the
affinities in the code itself.

Also we are clearing the virtual tx interrupt for all cpus
in case user changes the irq affinity to wrong cpu.

Signed-off-by: Tarun Garg <b10794@freescale.com>
Signed-off-by: Sandeep Gopalpet <sandeep.kumar@freescale.com>