gianfar: prevents user to change mtu when RPS is enabled
MTU change makes all skbs of a device to be destroyed
and re-allocated according to new mtu size. But the skbs
in cpu queues (used in RPS) cant be destroyed when mtu
of a device changes since cpu queue contains skbs from all
devices. If mtu of one of the device is increased and a
smaller size skb is recycled to this device from the cpu
queue, then the system can panic when a packet bigger
than skb size is received and put into this skb.

If MTU change is required along with RPS, then MTU of all
ETSEC devices should be changed to the new value before
RPS is enabled in the system.