tos: Replaces the tasklet implementation with
NAPI to make the system more balanced.

This patch updates the current tasklet implement to NAPI so as
the system is more balanced in the terms that the packet submission
and the packet forwarding after being processed can be done at
the same priority.

A per core dummy netdev structure has been used to make the NAPI work
on both the core for postprocessing.

Signed-off-by: Sandeep Malik <Sandeep.Malik@freescale.com>