Gianfar:Optimization for routed packets by exchanging rx/tx buffers
Patch enhances the Gianfar driver performance for the packets that are
received by the driver and transmitted by the driver in the same context.
Performance is enahnced by exchanging the skbs between the Tx and Rx rings,
rather than allocating new sbks from kernel or taking it from recycle queue.
The cleanup of BDs is done in the xmit funtions, so there is no need of the Tx intrupts,
or tx cleanup functionality in the receive path, avoing the locks for cleanups.
Patch gives best performance with DEFAULT_RX_RING_SIZE 32 DEFAULT_TX_RING_SIZE 32 and
GFAR_DEV_WEIGHT 32.
On P2020 with Packet offload Engine ASF, performace above 90% line rate is achieved for
opposite flow scenario, for single flow throughput is at line rate.

 Signed-off-by: Rajan Gupta <rajan.gupta@freescale.com>