powerpc/epapr: kernel panic in byte channel driver when probe fails
The byte channel driver registers byte channel devices with the TTY layer
during the OF probe.  If the probe fails (typically because the byte channel
nodes don't specify a TX interrupt), then the device is never registered.
This causes a kernel panic in the driver because the tty layer still tries
to the open the device, and the driver dereferences a null pointer.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>