powerpc/epapr: add a buffer in the byte chanel driver for tty output data
The ePAPR byte channel driver supports both the console and tty layers for
each byte channel.  The tty layer requires a .write_room function that
tells the layer how much free space is in the output buffer.  The tty
layer never sends more than this amount of data, so it expects the driver
to process every byte it receives from the layer.

Previously, the driver simply asked the hypervisor how much free space is
in its byte channel queue, but this caused a race condition.  During heavy
writes simultaneously from both the console and tty layers, the
hypervisor's byte channel queue has little available space.  If a console
write occurs after the .write_room function is called, but before the
corresponding tty write operation, then the hypervisor's byte channel
queue may not have enough room for all of the data from the tty layer.
This forces the driver to tell the tty layer that not all data has been
written.

To fix that, an internal buffer for the tty layer is added.  The
.write_room function returns the amount of free space is this buffer.  A
dequeue function is added to handle the actual writing of data from the
buffer to the byte channel.

Signed-off-by: Timur Tabi <timur@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>