Gianfar: Fix race conditions in skb recycling with pkt steering
There is a  race condition in skb recycling code when
it is used with packet steering:

Updation of global recycle queue: There are few places where
recycle queue spin lock is held while updating the link list
of skbs but counter for buffers in the queue is updated outside
the lock. So it may happen that buffer counter does not reflect
correct number of buffers in recycle queue.

Fix: Make updation of buffer counter and updation of list atomic
by putting them in spin lock.

Signed-off-by: Rajan Srivastava <rajan.srivastava@freescale.com>
Signed-off-by: Pankaj Chauhan <pankaj.chauhan@freescale.com>
Signed-off-by: akhil goyal <akhil.goyal@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>