fs/hugetlbfs/inode.c: Fix pgoff alignment checking
The statement:

vma->vm_pgoff & ~(huge_page_mask(h) >> PAGE_SHIFT)

is bogus as an alignment check on 32-bit powerpc. This
causes us to & the pgoff with something that looks like
this (for a 4M hugepage): 0xfff0_03ff. Instead, flip the
page size mask and then shift that, giving you 0x0000_03ff.

Signed-off-by: Becky Bruce <beckyb@kernel.crashing.org>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>