mpc85xx: Hugetlb for FSL BookE
Enable hugepages on Freescale BookE processors.  This allows the kernel to
use huge TLB entries to map pages, which can greatly reduce the number of
TLB misses and the amount of TLB thrashing experiences by applications with
large memory footprints.  Care should be taken when using this on FSL
processors, as the number of large TLB entries supported by the core is low
(16-64) on current processors.

Note that applications must either explicitly ask for hugepages or link
with libhugetlbfs and use environment variables to get hugepages - they
will not be allocated automatically.  This is intentional to allow the most
efficient use of the small number of large TLB entries supported by the
core.

The supported set of hugepage sizes include 4m, 16m, 64m, 256m, and 1g.
The numbers of larger pages sizes (64m+) that can be allocated in the
system may be limited and these page sizes should be used with care.  If
16m pages are being used, the kernel should be built with
CONFIG_FORCE_MAX_ZONE_ORDER=13.  Page sizes that are greater in size than
CONFIG_FORCE_MAX_ZONE_ORDER allows are considered "gigantic pages".  A
fixed number of these pages are allocated at boot time based on the command
line.  Pages that are within the size specified by
CONFIG_FORCE_MAX_ZONE_ORDER can be allocated/deallocated at any time when
the system is running.

Signed-off-by: Becky Bruce <beckyb@kernel.crashing.org>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>