SDHC: use mmc_delay instead of mdelay for time delay
mmc_delay() is a wrapper function for mdelay and msleep.
mdelay() blocks the system when busy-waiting.
msleep() suspends the currently running task to enable CPU to
process other tasks, so it is non-blocking regarding the whole
system.

When the desired delay time is more than a period of timer interrupt,
just use msleep().
Change mdelay to mmc_delay to avoid chewwing CPU when busy wait.

Signed-off-by: Shengzhou Liu <b36685@freescale.com>