dpaa_eth: Properly handle shared shared pools
If someone specifies a shared pool to be used by multiple interfaces,
the current code fails when it tries to register the same device twice.
The new code is more intelligent about detecting the sharing, and using
the existing buffer pool construct in dpa_bp_array[] whenever
interacting with buffer pools.

In addition, we remove support for buffer pools which have a "paddr" of
0. This was meant to serve as a mechanism for architects to specify pool
attributes, while letting the kernel allocate the memory, however:

1) The kernel now uses a shared pool by default, and allowing this
configuration just confuses things

2) The BMan driver will seed the pool with addresses, unless the size is
also 0, so this never was very useful.

Signed-off-by: Andy Fleming <afleming@freescale.com>