vlan: Reflect needed_headroom in vlan netdev
The stack checks needed_headroom before sending packets on to the
net_device, but when the device is on a VLAN, the VLAN code kept the
default.  This meant that drivers have to always check if there's enough
room, rendering needed_headroom useless.  We can fix this by copying
needed_headroom into the vlan device.

Signed-off-by: Andy Fleming <afleming@freescale.com>