powerpc/e5500: use a separate TLB handler when linear map is bolted
We don't need to worry about linear TLB misses, and if we do a full
table walk we get rid of all recursive TLB faults.  This gains a few
percent on TLB-miss-heavy workloads, and around 50% on a benchmark
that had a high rate of virtual page table faults under the normal
handler.

Signed-off-by: Scott Wood <scottwood@freescale.com>
Signed-off-by: Kumar Gala <galak@kernel.crashing.org>