Power Monitoring: ADD MCP3021 support
Add I2C driver for MCP3021 that is an ADC chip from Microchip.
On the MPC8536DS board, there are two MCP3021 chips to monitor the
core and platform current individually. The current is converted
to voltage through a current shunt monitor(INA196) and the voltage
then connected to MCP3021's analog input through a voltage-divider
network. On the board, the relationship between the core/platform
current and the MCP3021's input voltage is as follows:
Icore = 6.0 * Vin
Iplat = 5.5 * Vin
The driver export the value of Vin to sysfs, the voltage unit is
mV. Through the sysfs interface, lm-sensors tool can also display
Vin voltage.

The work voltage(Vdd) of the chip is passed into the driver through
dts file. Vdd is needed to convert the MCP3021's register value to
analog input voltage.

Signed-off-by: Mingkai Hu <Mingkai.hu@freescale.com>
Signed-off-by: Xie Xiaobo <r63061@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>