PIXIS: add board control FPGA support on MPC8536DS
The on-board FPGA supports an I2C IP to communicate with
the I2C device directly, such as temperature monitor chip
and two ADC chips which monitor the core and platform
current individually.

This function supports an ability to monitor the board
power consumption on the deep sleep mode as well as on
the normal mode, and these are what the patches do.

Signed-off-by: Mingkai Hu <Mingkai.hu@freescale.com>
Signed-off-by: Xie Xiaobo <r63061@freescale.com>
Signed-off-by: Li Yang <leoli@freescale.com>