powerpc/85xx: fsl-pamu: manually probe the PAMU nodes
The normal OF process calls the probe function at some indeterminate time
after the driver has loaded.  This is too late for us, because PAMU clients
(like the Qman driver) depend on PAMU being initialized early.

So instead, we "manually" call our probe function by creating the platform
devices ourselves.

In the process, we also add the assumption that there will only be one PAMU
node in the device tree.

Signed-off-by: Timur Tabi <timur@freescale.com>