qbman: remove recovery logic
The qbman drivers had an ad-hoc "recovery" scheme that needs to be
rethought. In particular it is assumes the use of a system-global
recovery of resources when the kernel boots, which provides no help to
implementing recovery in smaller-scale, more-dynamic use-cases, and
gets in the way of reworking portal initialisation for precisely the
same dynamic use-cases.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>