dpaa_eth: Increase FQ taildrop threshold
Raise the taildrop threshold for FQs, to avoid a large number of Tx
frames being dropped by QMan in 10G termination tests.

The new value is experimental, based on measurements from several
usecases. In the future we may want to make this a configuration option
and perhaps set different values for queues used by 1G and 10G ports.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>