dpaa_eth: Add extra headroom to Rx buffers
Add a configurable option for specifying if extra headroom is required
on the Rx path. This would instruct FMan to reserve some extra space
at the beginning of a data buffer, before Internal Context fields are
copied. This is in addition to the private data area already reserved
for driver internal use. It does not affect the buffer layout on the Tx
path

The value of the extra headroom can be tweaked from Kconfig or with a
bootarg (dpa_extra_headroom)

This feature is useful on particular forwarding scenarios that add extra
headers to the forwarded frame to avoid unbounded increase of recycled
buffers

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>
Signed-off-by: Madalin Bucur <madalin.bucur@freescale.com>
Signed-off-by: Stefan Szabo <szbs001@freescale.com>