dpaa_eth: Avoid recycling too large buffers
In certain conditions (forwarding with encapsulation and dpa extra
headroom set to zero), reallocated skbs get recycled leading to
the buffer pool being populated with increasingly large buffers.
This eventually leads to system wide memory depletion.

Add an extra check in the tx function to avoid recycling buffers
that are above a certain size.

Since we're here, also change the default value of the extra headroom
parameter from 0 to 64. This ensures best performance for forwarding
IPSec scenarios. For plain ipfwd or termination, using a value of
zero is recommended.

Signed-off-by: Ioana Radulescu <ruxandra.radulescu@freescale.com>