qman: set stash request queue to match the cpu affinity
Up until recently, the mapping of portals to stash request queues (and
the configuration of each stashing LIODN to target the desired cpu) was
being configured in u-boot using hard-coded constants. The device-tree
contained the same constants, and that's what linux was using to
determine which portals were used on which cpus.

Dynamic allocation of portals and dynamic assignment (and later,
reassignment) of cpu affinity has changed that, and the stash request
flow control feature was broken. The kernel needs to try to set the
portal's stash request queue at the moment it is configuring the portal
for cpu-affine use (ie. as it is allocated), and this is what this change
does. However this will fail with a warning if the kernel does not have
access to the QMan control registers.

Signed-off-by: Geoff Thorpe <Geoff.Thorpe@freescale.com>