dpaa_eth: Add CGR-based taildrop and congestion notification
Replace the FQTD mechanism with a more robust CGR-based congestion
detection. This allows us to notify the stack when the device is
congested, preventing it from pushing frames in a tight loop that would
waste CPU resources.
The actual congestion status-based taildrop mechanism, although enabled,
should be little-to-not used at runtime, if the kernel's egress queues
are timely stopped.

This change renounces the NETIF_F_HW_QDISC feature in the SG
driver. It remains still usable in the non-SG driver.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>