dpaa_eth: Differentiate Congestion State Notification thresholds
Define separate congestion thresholds on 1G and 10G egress ports,
because the former may become more easily flooded with egress traffic if
the cores are transmitting in a very tight loop, such as that of
netperf's UDP_STREAM tests.
If a threshold is too high and a device remains congested for too long,
there is a risk that the netdev watchdog fire and dump a disgraceful
call trace. Or we may run out of memory before that happens.
Conversely, if the threshold is too low, transmission will unjustifiedly
hiccup.

(To minimize the risk of a last-minute impact on the SDK,
we'll be leaving the two thresholds at the same value for now.
Situations such as the one described have not been reported in the SDK
releases yet, but this change is probably a good idea nevertheless.)

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>