dpaa_eth: Fix Rx frame validation logic
The frame validation logic on the Rx path was flawed in its assumptions
about the flags set by FMan's Hardware Parser. Specifically, an incoming
frame with bad L4 checksum (but otherwise properly encapsulated at L3
and L2) would go undetected and be passed as such on to the stack, which
would presumably have a hard time figuring out what to do with it.

Fix and refactor the frame validation logic, choose to drop such frames
which we might encounter, and update statistics to track this new kind
of event.

Signed-off-by: Bogdan Hamciuc <bogdan.hamciuc@freescale.com>