powerpc/pci: Remove the calling pcibios_setup_devices() in pcibios_enable_device()
For devices which are hot-plugged, when they are re-scanned,
only the probe function of hot-plugged devices will need to
request DMA and IRQ resources again. But for the operations
of PCIe AER, it is not necessary to request the DMA and IRQ
resources again. So remove the calling pcibios_setup_devices().

At the same time, use EXPORT_SYMBOL() to export pcibios_setup_device,
and let it as global function for other codes calling it.

Signed-off-by: Chunhe Lan <Chunhe.Lan@freescale.com>
Change-Id: I7eed6e08d5863414d179d2cee76308c0dce9a495
Reviewed-on: http://git.am.freescale.net:8181/6691
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Tiefei Zang <roy.zang@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>
(cherry picked from commit 424a80e8a957017794b23481e6a1af0fb94ad823)
Change-Id: I7eed6e08d5863414d179d2cee76308c0dce9a495
Reviewed-on: http://git.am.freescale.net:8181/7331
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>