dpa_offload: Improve memory allocation for statistics counters
The DPA Stats module is allocating a great deal of memory at initialization
time to be able to prepare for any counter or class counter the user
might create. As this was no longer acceptable with the addition of the
support for a maximum of 512 counters, the allocation was moved to counter
creation time. This allows the driver to allocate just enough memory for
each counter. As the counter type is known at creation time, the needed
memory space is greatly improved (reduced).

Change-Id: Ie13c5d1524cb68d0b3e61a21a8b1cea15f4397fa
Signed-off-by: Marian Chereji <marian.chereji@freescale.com>
Signed-off-by: Aurelian Zanoschi <Aurelian.Zanoschi@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/6708
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Anca Jeanina Floarea <anca.floarea@freescale.com>
Reviewed-by: Thomas Trefny <Tom.Trefny@freescale.com>
(cherry picked from commit ae9d2d8b31aca83a32be5c358d625d0a08dcc803)
Change-Id: Ie13c5d1524cb68d0b3e61a21a8b1cea15f4397fa
Reviewed-on: http://git.am.freescale.net:8181/7373
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>
Tested-by: Jose Rivera <German.Rivera@freescale.com>