powerpc/mpc85xx: fix 64bit system cannot up after hibernation resume
After bootcpu hibiernation restore the pages. we are now running
with the kernel data of the old kernel fully restored. if we reset
the non-bootcpus that will be reset cache(tlb), the non-bootcpus
will get new address(map virtual and physical address spaces).
but bootcpu tlb cache still use boot kernel data, so we need to
invalidate the bootcpu tlb cache make it to get new main memory data.

Signed-off-by: Wang Dongsheng <dongsheng.wang@freescale.com>
Change-Id: Id371f31566f4b07bdae545c02453e74471211159
Reviewed-on: http://git.am.freescale.net:8181/7376
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Chenhui Zhao <chenhui.zhao@freescale.com>
Reviewed-by: Yang Li <LeoLi@freescale.com>
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>