dpaa_eth: remove CONFIG_FSL_DPAA_TX_RECYCLE
The option to enable or disable Tx recycling for FMan v3
platforms was wrong - the recycling code is always assuming
that recycling is turned on and the config option only selected
the hw type. When running on FMan v3 with this option disabled
the non-FMan v3 recycling implementation was used erroneously.
Replacing the config option with the FMan v3 option.

Signed-off-by: Madalin Bucur <madalin.bucur@freescale.com>
Change-Id: Ide82bcf2679df6768ea926865d92daa5a1db0447
Reviewed-on: http://git.am.freescale.net:8181/8113
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Ruxandra Ioana Radulescu <ruxandra.radulescu@freescale.com>