powerpc/fsl-booke: Add T1040QDS L2 switch support and binding
T1040 SOC comes with an integrated L2 switching core with 8 ports of
1G and 2 ports of 2.5G. The L2 switch MDIO is wired to internal Lynx
PHYs. The two 2.5G ports are connected to the FMAN using fixed link
and SGMII connection. The 1G ports are connected to external PHY
through the SERDES using QSGMII connection. The addresses of the
internal PHY varies from 0x4 to 0xb on the MDIO bus.

Change-Id: I8ef1d1f1a256635f36fba57b0272a634af89229e
Signed-off-by: Stefan Sicleru <stefan.sicleru@freescale.com>
Signed-off-by: Alex Marginean <alexandru.marginean@freescale.com>
Signed-off-by: Razvan Stefanescu <razvan.stefanescu@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/8817
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Poonam Aggrwal <poonam.aggrwal@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>