powerpc: Add a vga alias node for T104x
In u-boot, when set the video as console, the name 'vga' is used
as a general name for the video device, during the fdt_fixup_stdout
process, the 'vga' name is used to search in the dtb to setup the
'linux,stdout-path' node. Though the T104x DIU is not VGA-compatible
device,to meet the 'vga' name used in u-boot, the vga alias node is added for
T104x in this patch. At the same time, a display alias is also added
so that no other components grow dependencies on the vga alias node.

Signed-off-by: Jason Jin <Jason.Jin@freescale.com>
Change-Id: I45a12332d870a017c7704d1633d0f2cc2a840218
Reviewed-on: http://git.am.freescale.net:8181/9167
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yang Li <LeoLi@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>