dpaa_eth: add print for proxy probe
There is no indication that the proxy interface successfully initialized the
attached MAC device. In this way, the user does not know if the device tree
is misconfigured or the MAC device successfully probed. Other ways to check
the MAC initialization exists, like MAC mem mapped registers, but a simple
message in the bootlog seems way more user-friendly.

Signed-off-by: Marian Rotariu <marian.rotariu@freescale.com>
Change-Id: I79b38719e6b49d8a4557b885a2b13854961f9b32
Reviewed-on: http://git.am.freescale.net:8181/9258
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Madalin-Cristian Bucur <madalin.bucur@freescale.com>
Reviewed-by: Jose Rivera <German.Rivera@freescale.com>